+++ example/iso_base_media_example.dart
@@ -1,42 +1,20 @@
 // ignore_for_file: avoid_print
 
-import 'dart:typed_data';
-
 import 'package:iso_base_media/iso_base_media.dart';
 
-Future<void> inspect() async {
-  final fileBox = await ISOSourceBox.openFilePath('./test/test_files/a.heic');
-  final s = await inspectISOBox(fileBox);
-  await fileBox.close();
-  print(s);
-}
-
-Future<void> extract() async {
-  final fileBox = await ISOSourceBox.openFilePath('./test/test_files/a.heic');
-  var s = '';
-  await inspectISOBox(fileBox, callback: (box, depth) async {
-    if (box.type == 'ispe') {
-      final data = await box.extractData();
-      s += '${uint8ListToHex(data)}\n';
-    }
-    return true;
-  });
-  await fileBox.close();
-  print(s);
-}
-
-String uint8ListToHex(Uint8List bytes) {
-  final StringBuffer buffer = StringBuffer();
-  buffer.write('bytes(${bytes.length}): ');
-  for (final byte in bytes) {
-    buffer.write(byte.toRadixString(16).padLeft(2, '0'));
-    buffer.write(' ');
+Future<void> main() async {
+  final fileBox = await ISOBox.openFileBoxFromPath('./test/test_files/a.heic');
+  final ipcoBox = await fileBox.getChildByTypePath(['meta', 'iprp', 'ipco']);
+  if (ipcoBox == null) {
+    print('ipco box not found');
+    return;
   }
-  return buffer.toString();
-}
-
-void main() async {
-  await inspect();
-  print('-----------------------');
-  await extract();
+  final hvcCBoxList = await ipcoBox.getDirectChildrenByTypes({'hvcC'});
+  if (hvcCBoxList.isEmpty) {
+    print('hvcC box not found');
+    return;
+  }
+  for (final hvcCBox in hvcCBoxList) {
+    print('hvcC: start: ${hvcCBox.headerOffset}, size: ${hvcCBox.boxSize}');
+  }
 }